backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
celery[redis]==5.4.0
redis==5.1.1
httpx==0.27.2
aiofiles==24.1.0
python-dateutil==2.9.0
pyyaml==6.0.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
